add ethereum dir
