+++ ethereum/contracts/utils/BytesUtils.sol
