+++ ethereum/deployer/Cargo.toml
@@ -1,0 +1,29 @@
+[package]
+name = "eth-deployer"
+version = "0.5.4"
+authors = ["LayerX Labs <div-labs@layerx.co.jp>"]
+edition = "2018"
+
+[dependencies]
+state-runtime-node-api = { path = "../../../nodes/state-runtime/api" }
+frame-common = { path = "../../../frame/common" }
+frame-sodium = { path = "../../../frame/sodium" }
+frame-config = { path = "../../../frame/config" }
+anonify-wallet = { path = "../../wallet" }
+reqwest = "0.9"
+clap = "~2.32"
+rand = "0.7"
+rand_core = "0.5"
+console = "0.7"
+dialoguer = "0.4"
+dirs = "2.0"
+log = "0.4"
+env_logger = "0.7"
+tiny-bip39 = "0.6.0"
+ed25519-dalek = "1.0.0-pre.2"
+web3 = "0.14"
+ethabi = "12.0.0"
+anyhow = "1.0"
+thiserror = "1.0"
+serde_json = "1.0"
+base64 = "0.11"
