+++ modules/anonify-eth-driver/src/eth/mod.rs
@@ -1,8 +1,6 @@
 pub mod connection;
-pub mod deployer;
 pub mod event_watcher;
 pub mod sender;
 
-pub use self::deployer::EthDeployer;
 pub use self::event_watcher::EventWatcher;
 pub use self::sender::EthSender;
